Fix bounds on table to make it threadsafe
To be clear on why I believe this is necessary: you call a hash() method on each key (which might modify it) without taking a lock (I believe) so your keys must be `Sync` for your type to be shared across threads; your iteration method returns & references to K and V so your values must be `Sync` as well.  The `Send` bound on the generic types is almost a given for anything multithreaded.

This is now the third project of three concurrent data structures I've seen that didn't realize this was necessary, so I'm worried that our documentation on this isn't very good :(  Can you propose any fixes that would make it clearer that you really, absolutely nearly always want to have bounds like this?